Make data batching sequential
Previosuly, random mini-batchs of data were taken, of 10 data points at
a time, using a random number generator. Now, the data is assumed to be
constructed of sequential tim series, with length `batch_size`. Other
changes:

- `plt.show()` added at the end of the training loop in `main`.
- State length and variance are now declared at the top of `main`,
instead of in the data module.
- A function has been added to the train module which creates a
generator of start and end indeces used for the training loop.